backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
py-clob-client==0.34.5
python-dotenv==1.0.1
httpx==0.28.1
websockets==14.1
pydantic==2.10.4
web3==7.6.0
